Year-1/Semester-2/Knowledge-Discovery-in-Text/Homework-01/requirements.txt
datasets==3.1.0
gensim==4.3.3
huggingface-hub==0.36.2
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
pyarrow==17.0.0
PyYAML==6.0.3
scikit-learn==1.3.2
scipy==1.10.1
torch==2.3.1
transformers==4.46.3
tqdm==4.67.3
